requirements.txt
asyncpg>=0.29.0
alembic>=1.10.4
authlib>=1.3.1
anyio>=4.6.0
certifi>=2024.8.30
fastapi>=0.115.0
fastapi_jwt>=0.3.0
google-api-python-client>=2.138.0
google-auth>=2.32.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1
google-cloud-language>=2.13.4
httpcore>=1.0.5
httpx>=0.27.2
sqlalchemy>=1.4.46
python-jose>=3.3.0
psycopg2-binary>=2.9.6
pydantic>=2.9.2
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9.6
pytest>=8.3.2
python-dotenv>=1.0.1
jinja2>=3.0.0
starlette>=0.38.6
typing_extensions>=4.12.2
redis>=5.0.8
uvicorn>=0.30.6



